[REFACTOR] Renaming of implementation class to SmartRecyclerAdapter and interface definition to ISmartRecyclerAdapter avoid casting to Recycler.Adapter.
