+++ smartadapter/src/main/java/smartadapter/SmartRecyclerAdapter.java
@@ -2,285 +2,346 @@ package smartadapter;
 
 /*
  * Created by Manne Öhlund on 2019-06-25.
- * Copyright (c) All rights reserved.
+ * Copyright © 2019 All rights reserved.
  */
 
 import android.support.annotation.NonNull;
 import android.support.v7.widget.RecyclerView;
+import android.view.ViewGroup;
 
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 
 import smartadapter.internal.Mapper;
+import smartadapter.listener.OnViewAttachedToWindowListener;
+import smartadapter.listener.OnViewDetachedFromWindowListener;
 import smartadapter.listener.ViewEventListener;
+import smartadapter.viewholder.RecyclableViewHolder;
 import smartadapter.viewholder.SmartViewHolder;
 import smartadapter.widget.ViewTypeResolver;
 
-public interface SmartRecyclerAdapter {
+/**
+ * SmartRecyclerAdapter is the core implementation of the library.
+ * It handles all the implementations of the {@link ISmartRecyclerAdapter} functionality.
+ */
+@SuppressWarnings("unchecked")
+public class SmartRecyclerAdapter extends RecyclerView.Adapter<SmartViewHolder> implements ISmartRecyclerAdapter {
+
+    private int itemCount = 0;
+    private List items = new ArrayList();
+
+    private final Mapper mapper;
+    private ViewTypeResolver viewTypeResolver;
+    private HashMap<Class<? extends SmartViewHolder>, HashMap<Integer, HashMap<Integer, ViewEventListener>>> viewEventListeners;
+
+    SmartRecyclerAdapter(Object callerEnclosingClass, List items) {
+        mapper = new Mapper(callerEnclosingClass);
+        setItems(items, false);
+        updateItemCount();
+    }
+
+    @Override
+    public int getItemViewType(int position) {
+        return mapper.getItemViewType(viewTypeResolver, items.get(position), position);
+    }
+
+    @NonNull
+    @Override
+    public SmartViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
+        return mapper.createViewHolder(viewEventListeners, parent, viewType);
+    }
+
+    @Override
+    public void onBindViewHolder(@NonNull SmartViewHolder holder, int position) {
+        holder.bind(items.get(position));
+    }
+
+    @Override
+    public void onViewRecycled(@NonNull SmartViewHolder holder) {
+        super.onViewRecycled(holder);
+        holder.unbind();
+    }
+
+    @Override
+    public boolean onFailedToRecycleView(@NonNull SmartViewHolder holder) {
+        if (holder instanceof RecyclableViewHolder) {
+            return ((RecyclableViewHolder)holder).onFailedToRecycleView();
+        }
+        return super.onFailedToRecycleView(holder);
+    }
+
+    @Override
+    public void onViewAttachedToWindow(@NonNull SmartViewHolder holder) {
+        super.onViewAttachedToWindow(holder);
+        if (holder instanceof OnViewAttachedToWindowListener) {
+            ((OnViewAttachedToWindowListener)holder).onViewAttachedToWindow();
+        }
+    }
+
+    @Override
+    public void onViewDetachedFromWindow(@NonNull SmartViewHolder holder) {
+        super.onViewDetachedFromWindow(holder);
+        if (holder instanceof OnViewDetachedFromWindowListener) {
+            ((OnViewDetachedFromWindowListener)holder).onViewDetachedFromWindow();
+        }
+    }
+
+    @Override
+    public int getItemCount() {
+        return itemCount;
+    }
+
+    @Override
+    public <T> int getItemCount(Class<T> type) {
+        int count = 0;
+        for (Object item : items) {
+            if (item.getClass().equals(type)) {
+                count++;
+            }
+        }
+        return count;
+    }
+
+    @Override
+    public Object getItem(int index) {
+        return items.get(index);
+    }
+
+    @Override
+    public List getItems() {
+        return items;
+    }
+
+    @Override
+    public <T> ArrayList<T> getItems(Class<T> type) {
+        ArrayList<T> itemOfType = new ArrayList<>();
+        for (Object item : items) {
+            if (item.getClass().equals(type)) {
+                itemOfType.add((T) item);
+            }
+        }
+        return itemOfType;
+    }
+
+    @Override
+    public void setItems(List items) {
+        setItems(items, true);
+    }
+
+    @Override
+    public void setItems(List items, boolean notifyDataSetChanged) {
+        if (items != null) {
+            this.items = items;
+            if (notifyDataSetChanged) {
+                smartNotifyDataSetChanged();
+            }
+        }
+    }
+
+    @Override
+    public void addItem(Object item) {
+        this.addItem(item, true);
+    }
+
+    @Override
+    public void addItem(Object item, boolean notifyDataSetChanged) {
+        if (item != null) {
+            this.items.add(item);
+            if (notifyDataSetChanged) {
+                smartNotifyDataSetChanged();
+            }
+        }
+    }
+
+    @Override
+    public void addItem(int index, Object item) {
+        addItem(index, item, true);
+    }
+
+    @Override
+    public void addItem(int index, Object item, boolean notifyDataSetChanged) {
+        if (item != null) {
+            this.items.add(index, item);
+            if (notifyDataSetChanged) {
+                smartNotifyItemInserted(index);
+            }
+        }
+    }
+
+    @Override
+    public void addItems(List items) {
+        this.addItems(items, true);
+    }
+
+    @Override
+    public void addItems(List items, boolean notifyDataSetChanged) {
+        if (items != null) {
+            this.items.addAll(items);
+            if (notifyDataSetChanged) {
+                smartNotifyDataSetChanged();
+            }
+        }
+    }
+
+    @Override
+    public void addItems(int index, List items) {
+        this.addItems(index, items, true);
+    }
+
+    @Override
+    public void addItems(int index, List items, boolean notifyDataSetChanged) {
+        if (items != null) {
+            this.items.addAll(index, items);
+            if (notifyDataSetChanged) {
+                smartNotifyDataSetChanged();
+            }
+        }
+    }
+
+    @Override
+    public boolean removeItem(int index) {
+        return this.removeItem(index, true);
+    }
+
+    @Override
+    public boolean removeItem(int index, boolean notifyDataSetChanged) {
+        if (items != null && !items.isEmpty()) {
+            this.items.remove(index);
+            if (notifyDataSetChanged) {
+                smartNotifyItemRemoved(index);
+            }
+            return true;
+        }
+        return false;
+    }
+
+    @Override
+    public boolean replaceItem(int index, Object item) {
+        return replaceItem(index, item, true);
+    }
+
+    @Override
+    public boolean replaceItem(int index, Object item, boolean notifyDataSetChanged) {
+        if (item != null) {
+            this.items.set(index, item);
+            if (notifyDataSetChanged) {
+                smartNotifyItemChanged(index);
+            }
+            return true;
+        }
+        return false;
+    }
+
+    @Override
+    public void clear() {
+        this.items.clear();
+        smartNotifyDataSetChanged();
+    }
+
+    @Override
+    public void smartNotifyDataSetChanged() {
+        updateItemCount();
+        notifyDataSetChanged();
+    }
+
+    @Override
+    public void smartNotifyItemChanged(int position) {
+        updateItemCount();
+        notifyItemChanged(position);
+    }
+
+    @Override
+    public void smartNotifyItemRangeChanged(int positionStart, int itemCount) {
+        updateItemCount();
+        notifyItemRangeChanged(positionStart, itemCount);
+    }
+
+    @Override
+    public void smartNotifyItemInserted(int position) {
+        updateItemCount();
+        notifyItemInserted(position);
+    }
+
+    @Override
+    public void smartNotifyItemRangeInserted(int positionStart, int itemCount) {
+        updateItemCount();
+        notifyItemRangeInserted(positionStart, itemCount);
+    }
+
+    @Override
+    public void smartNotifyItemRemoved(int position) {
+        updateItemCount();
+        notifyItemRemoved(position);
+    }
+
+    @Override
+    public void smartNotifyItemRangeRemoved(int positionStart, int itemCount) {
+        updateItemCount();
+        notifyItemRangeRemoved(positionStart, itemCount);
+    }
+
+    @Override
+    public final void updateItemCount() {
+        itemCount = items.size();
+    }
+
+    @Override
+    public void map(Class<?> itemType, Class<? extends SmartViewHolder> viewHolderType) {
+        mapper.addMapping(itemType, viewHolderType);
+    }
+
+    @Override
+    public Mapper getMapper() {
+        return mapper;
+    }
+
+    void setDataTypeViewHolderMapper(HashMap<String, Class<? extends SmartViewHolder>> dataTypeViewHolderMapper) {
+        mapper.setDataTypeViewHolderMapper(dataTypeViewHolderMapper);
+    }
+
+    void setSmartRecyclerAdapterMapper(HashMap<Class<? extends SmartViewHolder>, SmartRecyclerAdapter> smartRecyclerAdapterMapper) {
+        mapper.setSmartRecyclerAdapterMapper(smartRecyclerAdapterMapper);
+    }
+
+    @Override
+    public ViewTypeResolver getViewTypeResolver() {
+        return viewTypeResolver;
+    }
+
+    @Override
+    public void setViewTypeResolver(ViewTypeResolver viewTypeResolver) {
+        this.viewTypeResolver = viewTypeResolver;
+    }
+
+    @Override
+    public HashMap<Class<? extends SmartViewHolder>, HashMap<Integer, HashMap<Integer, ViewEventListener>>> getViewEventListeners() {
+        return this.viewEventListeners;
+    }
+
+    @Override
+    public HashMap<Integer, HashMap<Integer, ViewEventListener>> getViewEventListenersForViewHolder(Class<? extends SmartViewHolder> viewHolderType) {
+        return getViewEventListeners().get(viewHolderType);
+    }
+
+    @Override
+    public void setViewEventListeners(HashMap<Class<? extends SmartViewHolder>, HashMap<Integer, HashMap<Integer, ViewEventListener>>> viewEventListeners) {
+        this.viewEventListeners = viewEventListeners;
+    }
 
     /**
-     * Overrides {@link RecyclerView.Adapter#onViewDetachedFromWindow(RecyclerView.ViewHolder)}.
-     * @see RecyclerView.Adapter#onViewDetachedFromWindow(RecyclerView.ViewHolder)
-     */
-    void onViewDetachedFromWindow(@NonNull SmartViewHolder holder);
-
-    /**
-     * Overrides {@link RecyclerView.Adapter#getItemCount()}.
-     * @see RecyclerView.Adapter#getItemCount()
-     * @return data item count
-     */
-    int getItemCount();
-
-    /**
-     * Get item count for target class type.
-     * @param type target class type
-     * @param <T> type of class
-     * @return item count
-     */
-    <T> int getItemCount(Class<T> type);
-
-    /**
-     * Get item at index.
-     * @param index adapter index
-     * @return Data object for that index.
-     */
-    Object getItem(int index);
-
-    /**
-     * Get list of all data items.
-     * @return list of all data items
-     */
-    List getItems();
-
-    /**
-     * Get list of all data items for target class type.
-     * @param type target class type
-     * @param <T> type of class
-     * @return list of all data items for target class type
-     */
-    <T> ArrayList<T> getItems(Class<T> type);
-
-    /**
-     * Sets the data item list for the SmartRecyclerAdapter.
-     * Calls {@link #setItems(List, boolean)} with default notifyDataSetChanged to true.
-     * @param items list of data items
-     */
-    void setItems(List items);
-
-    /**
-     * Sets the data item list for the SmartRecyclerAdapter and notifies the RecyclerView to update.
-     * @param items list of data items
-     * @param notifyDataSetChanged indicates if RecyclerView should update.
-     */
-    void setItems(List items, boolean notifyDataSetChanged);
-
-    /**
-     * Adds item to the list of data.
-     * Calls {@link #addItem(Object, boolean)} with default notifyDataSetChanged to true.
-     * @param item any type of item
-     */
-    void addItem(Object item);
-
-    /**
-     * Adds item to the list of data and notifies the RecyclerView to update.
-     * @param item any type of item
-     * @param notifyDataSetChanged indicates if RecyclerView should update.
-     */
-    void addItem(Object item, boolean notifyDataSetChanged);
-
-    /**
-     * Adds item to the list of data at target index.
-     * Calls {@link #addItem(int, Object, boolean)} with default notifyDataSetChanged to true.
-     * @param index target index
-     * @param item any type of item
-     */
-    void addItem(int index, Object item);
-
-    /**
-     * Adds item to the list of data at target index and notifies the RecyclerView to update.
-     * @param index target index
-     * @param item any type of item
-     * @param notifyDataSetChanged indicates if RecyclerView should update.
-     */
-    void addItem(int index, Object item, boolean notifyDataSetChanged);
-
-    /**
-     * Adds a list of items to the SmartRecyclerAdapter list of data.
-     * Calls {@link #addItems(List, boolean)} with default notifyDataSetChanged to true.
-     * @param items list of items to add
-     */
-    void addItems(List items);
-
-    /**
-     * Adds a list of items to the SmartRecyclerAdapter list of data and notifies the RecyclerView to update.
-     * @param items list of items to add
-     * @param notifyDataSetChanged indicates if RecyclerView should update.
-     */
-    void addItems(List items, boolean notifyDataSetChanged);
-
-    /**
-     * Adds a list of items from index to the SmartRecyclerAdapter list of data.
-     * Calls {@link #addItems(List, boolean)} with default notifyDataSetChanged to true.
-     * @param index target index
-     * @param items list of items to add
-     */
-    void addItems(int index, List items);
-
-    /**
-     * Adds a list of items from index to the SmartRecyclerAdapter list of data and notifies the RecyclerView to update.
-     * @param index target index
-     * @param items list of items to add
-     * @param notifyDataSetChanged indicates if RecyclerView should update.
-     */
-    void addItems(int index, List items, boolean notifyDataSetChanged);
-
-    /**
-     * Removes item at index.
-     * @see #removeItem(int, boolean)
-     *
-     * @param index item index
-     * @return true if item was removed
-     */
-    boolean removeItem(int index);
-
-    /**
-     * Removes item at index.
-     * @param index item index
-     * @param notifyDataSetChanged updates recycler view with the new data
-     * @return true if item was removed
-     */
-    boolean removeItem(int index, boolean notifyDataSetChanged);
-
-    /**
-     * Replaces item at index.
-     * @see #replaceItem(int, Object, boolean)
-     *
-     * @param index item index
-     * @return true if item was replaced
-     */
-    boolean replaceItem(int index, Object item);
-
-    /**
-     * Replaces item at index.
-     * @param index item index
-     * @param notifyDataSetChanged updates recycler view with the new data
-     * @return true if item was replaced
-     */
-    boolean replaceItem(int index, Object item, boolean notifyDataSetChanged);
-
-    /**
-     * Clears all the data and calls {@link #smartNotifyDataSetChanged()}
-     */
-    void clear();
-
-    /**
-     * Calls {@link #updateItemCount()} and {@link RecyclerView.Adapter#notifyDataSetChanged()}
-     */
-    void smartNotifyDataSetChanged();
-
-    /**
-     * Notifies the recycler adapter that item at position has changed.
-     * Calls {@link #updateItemCount()} and {@link RecyclerView.Adapter#notifyItemChanged(int)}
-     * @param position adapter position.
-     */
-    void smartNotifyItemChanged(int position);
-
-    /**
-     * Notifies the recycler adapter that item range at position has changed.
-     * Calls {@link #updateItemCount()} and {@link RecyclerView.Adapter#notifyItemRangeChanged(int, int)}
-     * @param positionStart from position
-     * @param itemCount item count from positionStart
-     */
-    void smartNotifyItemRangeChanged(int positionStart, int itemCount);
-
-    /**
-     * Notifies the recycler adapter that item at position has been inserted.
-     * Calls {@link #updateItemCount()} and {@link RecyclerView.Adapter#notifyItemInserted(int)}
-     * @param position item inserted at this position
-     */
-    void smartNotifyItemInserted(int position);
-
-    /**
-     * Notifies the recycler adapter that item range from position has changed.
-     * Calls {@link #updateItemCount()} and {@link RecyclerView.Adapter#notifyItemRangeInserted(int, int)}
-     * @param positionStart from position
-     * @param itemCount item count from positionStart
-     */
-    void smartNotifyItemRangeInserted(int positionStart, int itemCount);
-
-    /**
-     * Notifies the recycler adapter that item at position has been removed.
-     * @param position item removed at this position
-     */
-    void smartNotifyItemRemoved(int position);
-
-    /**
-     * Notifies the recycler adapter that item range from position has been removed.
-     * Calls {@link #updateItemCount()} and {@link RecyclerView.Adapter#notifyItemRangeRemoved(int, int)}
-     * @param positionStart from position
-     * @param itemCount item count from positionStart
-     */
-    void smartNotifyItemRangeRemoved(int positionStart, int itemCount);
-
-    /**
-     * Updated the SmartRecyclerAdapter item count.
-     */
-    void updateItemCount();
-
-    /**
-     * Maps data item type with SmartViewHolder extension.
-     * @param itemType data item type
-     * @param viewHolderType view holder type
-     */
-    void map(Class<?> itemType, Class<? extends SmartViewHolder> viewHolderType);
-
-    /**
-     * Returns the data item view holder mapper.
-     * @return Mapper
-     */
-    Mapper getMapper();
-
-    /**
-     * Returns {@link ViewTypeResolver}.
-     * @return viewTypeResolver
-     */
-    ViewTypeResolver getViewTypeResolver();
-
-    /**
-     * Sets {@link ViewTypeResolver}.
-     * @param viewTypeResolver the ViewTypeResolver
-     */
-    void setViewTypeResolver(ViewTypeResolver viewTypeResolver);
-
-    /**
-     * Get all ViewEventListeners.
-     * @return map of ViewEventListeners
-     */
-    HashMap<Class<? extends SmartViewHolder>, HashMap<Integer, HashMap<Integer, ViewEventListener>>> getViewEventListeners();
-
-    /**
-     * Get ViewEventListeners for target ViewHolder.
-     * @param viewHolderType SmartViewHolder type
-     * @return map of ViewEventListeners
-     */
-    HashMap<Integer, HashMap<Integer, ViewEventListener>> getViewEventListenersForViewHolder(Class<? extends SmartViewHolder> viewHolderType);
-
-    /**
-     * Sets map of {@link ViewTypeResolver}.
-     * @param viewEventListeners map of ViewEventListeners
-     */
-    void setViewEventListeners(HashMap<Class<? extends SmartViewHolder>, HashMap<Integer, HashMap<Integer, ViewEventListener>>> viewEventListeners);
-
-    /**
-     * Builder of {@link SmartRecyclerAdapter} for easy implementation
+     * Builder of {@link SmartRecyclerAdapter} for easy implementation.
      * @return SmartAdapterBuilder
      */
-    static SmartAdapterBuilder items(List items) {
+    public static SmartAdapterBuilder items(List items) {
         return new SmartAdapterBuilder(items);
     }
 
-    static SmartAdapterBuilder empty() {
+    /**
+     * Builder of {@link SmartRecyclerAdapter} for easy implementation.
+     * @return SmartAdapterBuilder
+     */
+    public static SmartAdapterBuilder empty() {
         return new SmartAdapterBuilder(new ArrayList());
     }
 }
